+++ src/tui/mod.rs
@@ -11,6 +11,7 @@ mod dict;
 pub mod events;
 pub mod git;
 pub mod hackmd_meta;
+mod images;
 mod jsonl;
 mod links;
 mod markdown;